Fix typing client sample, sender and updater

+++ misc/client_sample.py
@@ -1,6 +1,6 @@
 from PyQt6 import QtWidgets, uic
 from PyQt6.QtCore import pyqtSignal, QThreadPool, QEvent, QObject
-from PyQt6.QtWidgets import QPushButton, QLineEdit, QProgressBar, QTextEdit, QLabel, QStackedWidget, QSpinBox, QSlider, QDoubleSpinBox
+from PyQt6.QtWidgets import QPushButton, QLineEdit, QProgressBar, QTextEdit, QLabel, QStackedWidget, QSpinBox, QSlider, QDoubleSpinBox, QWidget
 
 from src.core.config import Config
 from src.utils.constants import constants
@@ -19,7 +19,7 @@ def resource_path(relative_path):
     """ Retorna o caminho absoluto para o recurso, compatível com PyInstaller """
     if hasattr(sys, '_MEIPASS'):
         # No executável, sys._MEIPASS é a raiz da pasta temporária
-        base_path = sys._MEIPASS
+        base_path = getattr(sys, '_MEIPASS', os.path.dirname(sys.executable))
     else:
         # No desenvolvimento, a base é a pasta raiz do projeto (onde está o main4.py)
         # Como este arquivo está em misc, pegamos o pai dele
@@ -39,9 +39,9 @@ class ClientSimulator(QtWidgets.QMainWindow):
     _client_transaction_ID = 0
     _client_name = "Simulator"
 
-    def __init__(self, clientID: int=None, clientName: str=None):
+    def __init__(self, clientID: int | None=None, clientName: str | None=None):
         super(ClientSimulator, self).__init__()
-        uic.loadUi(main_ui_path, self)
+        uic.loadUi(main_ui_path, self) # type: ignore
 
         if not self._check_config():
             return
@@ -108,7 +108,11 @@ class ClientSimulator(QtWidgets.QMainWindow):
         self.pageSelect: QStackedWidget = self.pageSelect
 
         self.stsBar = self.findChild(QtWidgets.QStatusBar, 'stsBar')
-        self.stsBar: QStackedWidget = self.stsBar
+        self.stsBar: QtWidgets.QStatusBar = self.stsBar
+
+        sb = self.statusBar()
+        if sb is not None:
+            self.status_bar: QtWidgets.QStatusBar = sb
 
         self.sbMovePos: QDoubleSpinBox = self.findChild(QDoubleSpinBox, "sbMovePos")
 
@@ -280,7 +284,7 @@ class ClientSimulator(QtWidgets.QMainWindow):
             self.lblClientID.setText(str(self.client_ID))
             self.sbMovePos.setValue(data[SJson.MAX_STEP])
             
-            self.statusBar().clearMessage()
+            self.status_bar.clearMessage()
             self.pageSelect.setCurrentIndex(1)
 
             self.threadpool.start(self._updater)    # Starts updater
@@ -289,7 +293,7 @@ class ClientSimulator(QtWidgets.QMainWindow):
 
         except Exception as e:
             print({str(e)})
-            self.statusBar().showMessage("Could not establish connection to server")
+            self.status_bar.showMessage("Could not establish connection to server")
 
     def _clientIpDefined(self):
         self.btnConnectClient.click()
@@ -304,39 +308,44 @@ class ClientSimulator(QtWidgets.QMainWindow):
             return False
 
     def _start_client(self):
-        self.txtStatus.setText("Connecting subscriber socket...")
-        self.subscriber = self.context.socket(zmq.SUB)
-        self.subscriber.connect(f"tcp://{self._connection_ip}:{Config.port_pub}")
-        topics_to_subscribe = ''
+        if self.context is not None:
+            self.txtStatus.setText("Connecting subscriber socket...")
+            self.subscriber = self.context.socket(zmq.SUB)
+            self.subscriber.connect(f"tcp://{self._connection_ip}:{Config.port_pub}")
+            topics_to_subscribe = ''
 
-        self.subscriber.setsockopt_string(zmq.SUBSCRIBE, topics_to_subscribe)
+            self.subscriber.setsockopt_string(zmq.SUBSCRIBE, topics_to_subscribe)
 
-        self.poller = zmq.Poller()
-        self.poller.register(self.subscriber, zmq.POLLIN)
+            self.poller = zmq.Poller()
+            self.poller.register(self.subscriber, zmq.POLLIN)
 
-        self.txtStatus.setText("Connecting requisition socket...")
-        self.req = self.context.socket(zmq.REQ)
-        self.req.connect(f"tcp://{self._connection_ip}:{Config.port_rep}")
-        self.txtStatus.clear()
+            self.txtStatus.setText("Connecting requisition socket...")
+            self.req = self.context.socket(zmq.REQ)
+            self.req.connect(f"tcp://{self._connection_ip}:{Config.port_rep}")
+            self.txtStatus.clear()
        
     def _reset_client_context(self):
         """ Resets client context to allow continuous communication in case of a timeout  """
         try:
-            self._clear_thread_updater()
-            self._clear_thread_sender()
-            self.context.destroy()
-            self._connect_to_server()
-            print(f"[ZMQ Client] Communication reset success")
+            if self.context is not None:
+                self._clear_thread_updater()
+                self._clear_thread_sender()
+                self.context.destroy()
+                self._connect_to_server()
+                print(f"[ZMQ Client] Communication reset success")
         except:
             print(f"[ZMQ Client] Error establishing connection")
         
     
     def _send_command(self, command: str, timeout: int=5000) -> str: #1500
         try:
-            self.transaction_ID += 1                                        #   Updates transaction ID
-            self._sender.send_request(self, command, timeout)               #   Sets message
-            self.threadpool.start(self._sender)                             #   Starts Sender thread    
-            return "OK"
+            if self._sender is not None:
+                self.transaction_ID += 1                                        #   Updates transaction ID
+                self._sender.send_request(self, command, timeout)               #   Sets message
+                self.threadpool.start(self._sender)                             #   Starts Sender thread    
+                return "OK"
+            else:
+                raise Exception("Sender thread is not initialized. Please connect to the server first.")
         except Exception as e:
             return f"[ZMQ Client] Error sending command to server -> {str(e)}"
 
@@ -369,30 +378,31 @@ class ClientSimulator(QtWidgets.QMainWindow):
 
     def _move_in(self):
         # status = json.loads(self.txtStatus.toPlainText())
-
-        max_speed = int(self._updater.data['maxSpeed'])                 # actually is the Normal speed configuration
-        if not self.is_moving:
-            if TEST_SETUP:
-                self._send_command(f"FOCUSIN=" + f"{str(int(max_speed * self.sbSpeed.value() * 0.01))}")   #TEST_VALUE -> ORIGINAL VALUE => FOCUSIN=200
+        if self._updater is not None:
+            max_speed = int(self._updater.data[SJson.MAX_SPEED])  #  actually is the Normal speed configuration
+            if not self.is_moving:
+                if TEST_SETUP:
+                    self._send_command(f"FOCUSIN=" + f"{str(int(max_speed * self.sbSpeed.value() * 0.01))}")   #TEST_VALUE -> ORIGINAL VALUE => FOCUSIN=200
+                else:
+                    self._send_command(f"FOCUSIN=" + f"{str(int(max_speed * self.sbSpeed.value() * 0.01))}")
             else:
-                self._send_command(f"FOCUSIN=" + f"{str(int(max_speed * self.sbSpeed.value() * 0.01))}")
-        else:
-            self._send_command("HALT")
+                self._send_command("HALT")
 
     def _move_out(self):
         # status = json.loads(self.txtStatus.toPlainText())
         # max_speed = int(status['maxSpeed'])                 # actually is the Normal speed configuration
-        max_speed = int(self._updater.data['maxSpeed'])
-        if not self.is_moving:
-            if TEST_SETUP:
-                self._send_command(f"FOCUSOUT="  + f"{str(int(max_speed * self.sbSpeed.value() * 0.01))}")  #TEST_VALUE -> ORIGINAL VALUE => FOCUSOUT=200
+        if self._updater is not None:
+            max_speed = int(self._updater.data[SJson.MAX_SPEED])  #  actually is the Normal speed configuration
+            if not self.is_moving:
+                if TEST_SETUP:
+                    self._send_command(f"FOCUSOUT="  + f"{str(int(max_speed * self.sbSpeed.value() * 0.01))}")  #TEST_VALUE -> ORIGINAL VALUE => FOCUSOUT=200
+                else:
+                    self._send_command(f"FOCUSOUT="  + f"{str(int(max_speed * self.sbSpeed.value() * 0.01))}")
             else:
-                self._send_command(f"FOCUSOUT="  + f"{str(int(max_speed * self.sbSpeed.value() * 0.01))}")
-        else:
-            self._send_command("HALT")
+                self._send_command("HALT")
                 
     def _get_status(self):
-        if self._sender._send is False:                          # A new command is only sent if the last one was already sent
+        if self._sender and self._sender._send is False:                          # A new command is only sent if the last one was already sent
             self._send_command("STATUS")
 
             
@@ -421,20 +431,24 @@ class ClientSimulator(QtWidgets.QMainWindow):
         self.is_moving = status
 
 
-    def _update_gui_element(self, widget: QtWidgets):
-        """Updates the GUI element style after an event occured
-
+    def _update_gui_element(self, widget: QWidget):
+        """Updates the GUI element style after an event occured.
+        According to QT framework this functions must be executed to update visual elements when a property is changed.
+        Re-polish the style to apply CSS changes linked to this property
+        
         Parameters
         ----------
         widget : QtWidgets
             Widget to be updated
         """
-        widget.style().unpolish(widget)
-        widget.style().polish(widget)
+        w_style = widget.style()
+        if w_style:
+            w_style.unpolish(widget)
+            w_style.polish(widget)
         widget.update()
 
 
-    def eventFilter(self, obj: QObject, event: QEvent) -> bool:
+    def eventFilter(self, obj: QWidget, event: QEvent) -> bool:
         """Process events
 
         Parameters
